First checkin of kube-up support for photon-controller
This supports kube-up as it exists on the release-1.2 branch (not
master). It supports kube-up and kube-down, but does not yet support
kube-push or the integrated tests.

Manual tests work well: we can bring up a cluster, create and
manipulate pods, and bring down the cluster.

Networking to public IPs for services has not been tested. That will
be tested next, and integrated tests will follow shortly thereafter.

There is an extra script called setup-prereqs.sh. This configures
Photon Controller with tenant, project, flavors, and image. It's
unlikely to be useful for production deployments because it is an
opinionated configuration, but it will be useful for getting people up
and running quickly.