Fix: Pod IP addresses were not in correct IP range.
Docker didn't start up correctly on the minions because we started
installed our custom docker.service file before docker. Our custom
file was overwritten and not used, and therefore the correct IPs
weren't used.

Debugging this turned up a minor error in how we log during the
installation of salt, so taht was also fixed.