Fix setup of pod routes
There was a race condition in setting up the pod routes that could
result in them not working.

When we set up the routes, we wait for the Linux bridge (cbr0) to have
an IP address. We then find the subnet it's on (e.g. 10.244.0.0/24)
and set up routes based on that subnet.

However, when the bridge comes up, it comes up in stages, and we'll
see two IP addresses. The first IP address is in the 172.x.x.x
network, and the second one will be the real network. The code would
sometimes use the 172 network and sometimes the real network.

This update waits for the real address, and now pod networking works:
all nodes can access all pod and cluster IP addresses.