Removed the use of sshpass
While bootstrapping this code, we used a VMDK with a well-known
username/password and sshpass. While we did install ssh keys and use
them, the initial bootstrap relied on that password.

The kube image has been updated to remove that account and add
cloud-init. When we start up the VM, we make an ISO that contains the
configuration for cloud-init (the user-data.txt file). This tells
cloud-init to make the kube account. It will be configured with the
ssh key we were already using. We do give it a password (backup login)
using the password randomly generated by kube-up.