Fix use of awk to parse output of photon command
Awk on Mac and Linux are slightly different. We want awk to split
fields on tabs (not all whitespace, just tabs, so we can handle spaces
in fields). On the Mac there is a shortcut to specify tabs as field
separators using -Ft. That's apparently not portable, so this changes
it to -F'\t'.